+++ ExposurePlayback/Context/Source/AssetSource.swift
@@ -17,7 +17,7 @@ import Player
 /// * `.customPosition(position:)` : Playback starts from the specified buffer position (in milliseconds) . Will ignore positions outside the `seekableRange`.
 /// * `.customTime(time:)` : Starting from a unix timestamp is undefined for Vod assets. Will use `.defaultBehavior`
 /// * `.defaultBehavior` Playback starts from the beginning of the asset
-public class AssetSource: ExposureSource { }
+open class AssetSource: ExposureSource { }
 
 extension AssetSource: ContextPositionSeekable {
     internal func handleSeek(toPosition position: Int64, for player: Player<HLSNative<ExposureContext>>, in context: ExposureContext) {