+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 #### Features
 * Added *Exposure* `X-Request-Id` to license requests.
 * `EMP-11711` Playtoken passed as header when requesting fairplay licenses.
+* `EMP-11747` Added `EntitlementSourceResponseHeaders` protocol.
 
 #### Changes
 * Stop and unload `MediaAsset` when playback reaches end of duration.